+++ Buildenator.Abstraction/MakeBuilderAttribute.cs
@@ -1,8 +1,9 @@
 ﻿using System;
+// ReSharper disable UnusedParameter.Local
 
 namespace Buildenator.Abstraction
 {
-    [AttributeUsage(AttributeTargets.Class, Inherited = false, AllowMultiple = false)]
+    [AttributeUsage(AttributeTargets.Class, Inherited = false)]
     public sealed class MakeBuilderAttribute : Attribute
     {
         /// <summary>
@@ -12,12 +13,13 @@ namespace Buildenator.Abstraction
         /// <param name="buildingMethodsPrefix">How the builder methods should be named.</param>
         /// <param name="defaultStaticCreator">The resulting builder will have a special static building method with default parameters. true/false/null</param>
         /// <param name="nullableStrategy">Change nullable context behaviour. Use the <see cref="NullableStrategy"/> enum.</param>
+        /// <param name="generateMethodsForUnreachableProperties"></param>
         public MakeBuilderAttribute(
             Type typeForBuilder,
             string? buildingMethodsPrefix = "With",
             object? defaultStaticCreator = null,
             object? nullableStrategy = null,
-            object? generateMethodsForUnrechableProperties = null
+            object? generateMethodsForUnreachableProperties = null
             )
         {
         }