+++ Tests/Buildenator.IntegrationTests.Source/Builders/DefaultConstructorBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using Buildenator.Abstraction;
 using Buildenator.IntegrationTests.SharedEntities;
 
-namespace Buildenator.IntegrationTests.Builders
+namespace Buildenator.IntegrationTests.Source.Builders
 {
     [MakeBuilder(typeof(DefaultConstructor), defaultStaticCreator: false)]
     public partial class DefaultConstructorBuilder