+++ Tests/Buildenator.IntegrationTests.SourceWithoutAssemblyInfo/DifferentNameBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using Buildenator.Abstraction;
 using Buildenator.IntegrationTests.SharedEntities;
 
-namespace Buildenator.IntegrationTests.Source.Builders
+namespace Buildenator.IntegrationTests.SourceWithoutAssemblyInfo
 {
     [MakeBuilder(typeof(Entity))]
     public partial class DifferentNameBuilder